CCC ON THE PROTOCOLS FOR TOP-LEVEL VIEWS
Code Cleanup;
Well this is some nice news, I was able to sort out the protocol and
will be MD'ing them shortly (as to whether they are incoming or
outgoing interface) And I also changed the names to be more sensible.
Like the Older Cocoa API. I could almost add the correct implementation
of Sessions here. This is a fix of almost the right version from
yesterday